added generator for iterating over all class ids
this is a common usecase, so we provide a convenient way to do it. The
generator returns one class for every class id. As many different
class instances can exist for every id, one should be careful only to
use class properties when using this iterator, not e.g. properties of
one of the standards of this class.